+++ application/core/MY_Model.php
@@ -487,6 +487,7 @@ include_once 'Entity.php'; // Entity class
 include_once 'DataMapper.php'; // Data mapper
 include_once 'Memory_Model.php';	// In-memory only
 include_once 'CSV_Model.php';	// CSV persisted
-
+include_once 'Model_Entity_Controller_Helper.php'; // Interface for models to implement helper functions for controllers
+include_once 'Model_Controller_Helper.php'; // Class for csv_models to make it easier for controllers
 
 /* End of file */